+++ js/data.js
@@ -1,0 +1,4 @@
+import './utils/create-photo-descriptions.js';
+import {createPhotoDescriptions, USER_PHOTO_DESCRIPTIONS} from './utils/create-photo-descriptions.js';
+
+createPhotoDescriptions(USER_PHOTO_DESCRIPTIONS);
